+++ ObjectPrinting/PropertySerializingConfigExtensions.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Globalization;
 using System.Linq;
-using System.Linq.Expressions;
 
 namespace ObjectPrinting
 {
